+++ api/types/image/opts.go
@@ -106,6 +106,15 @@ type LoadOptions struct {
 type InspectOptions struct {
 	// Manifests returns the image manifests.
 	Manifests bool
+
+	// Platform selects the specific platform of a multi-platform image to inspect.
+	//
+	// If the server is not multi-platform capable, this option will make the
+	// server return an error if the actual image platform doesn't match this
+	// platform.
+	//
+	// This option is only available for API version 1.49 and up.
+	Platform *ocispec.Platform
 }
 
 // SaveOptions holds parameters to save images.